create clear data page and move functionality from calendar page to clear data page
