+++ pages/settings/data/clear-data.js
@@ -1,0 +1,66 @@
+import { useState } from 'react';
+
+import Snackbar from '../../../shared/components/Snackbar/Snackbar';
+
+export default function ClearData () {
+    const [ snackbars, setSnackbars ] = useState({
+        clear: {
+            snackbar: null,
+            timeout: null
+        }
+    });
+
+    const clearLocalStorage = () => {
+        let cloneSnackbars = JSON.parse(JSON.stringify(snackbars));
+        if (cloneSnackbars.clear.timeout) {
+            clearTimeout(snackbars.clear.timeout);
+        }
+
+        let message = 'cleared all data';
+        cloneSnackbars.clear.snackbar = {
+            message: message,
+            className: 'snackbar-red'
+        }
+
+        let snackbarTimeout = setTimeout(() => {
+            let cloneSnackbars = JSON.parse(JSON.stringify(snackbars));
+            cloneSnackbars.clear.snackbar = null;
+            cloneSnackbars.clear.timeout = null;
+            setSnackbars(previousSnackbars => { 
+                return {
+                    ...previousSnackbars,
+                    clear: cloneSnackbars.clear
+                }
+            });
+        }, 5000);
+
+        cloneSnackbars.clear.timeout = snackbarTimeout;
+        setSnackbars(cloneSnackbars);
+
+        localStorage.clear();
+    }
+
+    return (
+        <div className="container">
+            <div className="mx-15">
+                <h1>
+                    clear data 
+                </h1>
+
+                <p>
+                    clear food blocks and previous food blocks
+                </p>
+
+                <button
+                    onClick={ clearLocalStorage }
+                    className="button button-red flex-grow-1">clear data</button>
+
+                <div className="snackbars-container">
+                    { snackbars.clear.snackbar ? (
+                        <Snackbar message={ snackbars.clear.snackbar.message } className={ snackbars.clear.snackbar.className } />
+                    ) : (null) }
+                </div>
+            </div>
+        </div>
+    );
+}
